CGFloat换成float以适配

+++ WTRequestCenter/UIKit+WTRequestCenter/WTDataSaver.h
@@ -62,7 +62,7 @@
 //Available in iOS 2.0 and later.
 
 
-+(CGFloat)osVersion;
++(float)osVersion;
 
 #pragma mark - 对象转换
 //用于对象和Data的转换，类型可以是NSDictionary或者NSArray
